refactor: move gemini cli onboarding helpers
